moved WCPMTInfo struct out of NuSelectionBDT
